# src/main/java/com/sp/addressbook/model/Name.java
package com.sp.addressbook.model;

import com.fasterxml.jackson.annotation.JsonAutoDetect;

import java.util.Objects;

@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
public class Name {

    private String firstName;
    private String lastName;

    public Name(String firstName, String lastName) {
        this.firstName = firstName;
        this.lastName = lastName;
    }

    public Name() {}

    public String getFirstName() {
        return firstName;
    }

    public String getLastName() {
        return lastName;
    }

    @Override
    public int hashCode() {
        return Objects.hash(this.firstName, this.lastName);
    }

    @Override
    public boolean equals(Object obj) {
        if (obj == this)
            return true;
        if (obj == null || !(obj instanceof Name))
            return false;
        Name name = (Name) obj;
        return name.getFirstName().equals(this.getFirstName()) && name.getLastName().equals(this.lastName);
    }

    @Override
    public String toString() {
        return "Name : " + this.firstName + " " + this.lastName;
    }

}